Initial extracted base module
