+++ app/build.gradle
@@ -46,23 +46,13 @@ android {
 }
 
 dependencies {
+    implementation project(path: ':base')
     //region implementation dependencies
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:${versions.kotlin}"
 
     implementation "com.google.dagger:hilt-android:${versions.hilt}"
     implementation "androidx.hilt:hilt-lifecycle-viewmodel:${versions.hiltx}"
-
-    implementation "io.reactivex.rxjava3:rxjava:${versions.rxJava}"
-    implementation "io.reactivex.rxjava3:rxandroid:${versions.rxAndroid}"
-    implementation "io.reactivex.rxjava3:rxkotlin:${versions.rxKotlin}"
-    implementation "com.github.akarnokd:rxjava3-extensions:${versions.rxExt}"
-
-    implementation "com.jakewharton.rxbinding4:rxbinding:${versions.rxBinding}"
-    implementation "com.jakewharton.rxbinding4:rxbinding-core:${versions.rxBinding}"
-    implementation "com.jakewharton.rxbinding4:rxbinding-appcompat:${versions.rxBinding}"
-    implementation "com.jakewharton.rxbinding4:rxbinding-recyclerview:${versions.rxBinding}"
-    implementation "com.jakewharton.rxbinding4:rxbinding-material:${versions.rxBinding}"
 
     implementation "com.google.android.material:material:${versions.material}"
     implementation "org.jetbrains.anko:anko-commons:${versions.anko}"
