+++ app/src/main/java/com/example/appName/presentation/features/main/MainPresenter.kt
@@ -2,7 +2,7 @@ package com.example.appName.presentation.features.main
 
 import androidx.hilt.lifecycle.ViewModelInject
 import com.example.appName.data.repository.exampleuser.ExampleUserRepository
-import com.example.appName.presentation.features.base.BasePresenter
+import com.example.base.BasePresenter
 import com.example.appName.presentation.features.main.MainViewState.PartialState.LoggedInState
 import com.example.appName.presentation.features.main.MainViewState.PartialState.WelcomeState
 import io.reactivex.rxjava3.core.Flowable
