+++ base/src/test/java/com/example/base/BasePresenterTest.kt
@@ -1,6 +1,7 @@
-package com.example.appName.presentation.features.base
+package com.example.base
 
-import com.example.appName.utils.TestSchedulersFactory
+import com.example.base.utils.SchedulersFactory
+import com.example.base.testutils.TestSchedulersFactory
 import io.reactivex.rxjava3.core.Flowable
 import org.junit.Before
 import org.junit.Test
