+++ build.gradle
@@ -2,6 +2,9 @@ apply plugin: 'com.github.ben-manes.versions'
 
 buildscript {
     apply from: 'config.gradle'
+    ext {
+        kotlin_version = '1.3.72'
+    }
 
     repositories {
         google()
